+++ GUI/AchievementTooltip/MostProgress.lua
@@ -42,18 +42,20 @@ local function GetDetails(character, achievementId)
 	if not achievementProgress then
 		return;
 	end
-
+	local achievementNumCriteria = GetAchievementNumCriteria(achievementId);
 	local details = {};
 	for i, progress in next, achievementProgress do
-		local criteriaString, _, completed, quantity, reqQuantity, _, _, _, _, _, _, hasValueProgress = addon.GetAchievementCriteriaInfo(achievementId, i);
-		if type(progress) == "boolean" and progress then
-			completed = true;
-			quantity = reqQuantity;
-		elseif type(progress) == "number" then
-			completed = progress >= reqQuantity;
-			quantity = progress;
+		if i <= achievementNumCriteria then
+			local criteriaString, _, completed, quantity, reqQuantity, _, _, _, _, _, _, hasValueProgress = addon.GetAchievementCriteriaInfo(achievementId, i);
+			if type(progress) == "boolean" and progress then
+				completed = true;
+				quantity = reqQuantity;
+			elseif type(progress) == "number" then
+				completed = progress >= reqQuantity;
+				quantity = progress;
+			end
+			tinsert(details, {criteriaString, completed, quantity, reqQuantity, hasValueProgress});
 		end
-		tinsert(details, {criteriaString, completed, quantity, reqQuantity, hasValueProgress});
 	end
 	return details;
 end