+++ Localization/enUS.lua
@@ -1,4 +1,4 @@
--- [[ Exported at 2023-03-23 07-53-40 ]] --
+-- [[ Exported at 2023-03-23 19-45-49 ]] --
 -- [[ This code is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. ]] --
 
@@ -614,4 +614,6 @@ L["Ignore Filters"] = "Ignore Filters";
 L["Ignore Filters Desc"] = "Ignore any filters and show all the achievements in the {category}.";
 L["until the end of"] = "until the end of";
 L["Always Show Completed"] = "Always Show Completed";
+L["Show Faction Faction Icon"] = "Show {faction} Faction Icon";
+L["Show Faction Faction Icon Desc"] = "Show the {faction} faction icon on the achievement if only {faction} players can earn it.";