+++ Localization/zhCN.ExtraWrathStrings.lua
@@ -1,4 +1,4 @@
--- [[ Exported at 2023-03-23 07-53-40 ]] --
+-- [[ Exported at 2023-03-23 19-45-49 ]] --
 -- [[ This code is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. ]] --