+++ CHANGELOG.md
@@ -7,7 +7,7 @@ All notable changes to this project will be documented in this file.
 
 ### Fixed
 - Removed a ton of achievements from the new zone
-- Workaround (plugin) for Wowhead Quick Link since the author is not active right now (issue #105)
+- Rare error when an achievement of the opposing faction has more criteria than is visible by the current faction
 
 ## 54.1 - 2023-03-22
 ### Fixed